Handle empty blocks when decompressing DEFLATE streams.
Trying to read an empty block triggers an error when reaching into the
block to copy it to the output buffer.  Since some implementations, like
Go, will generate empty static blocks as a way to end an output stream,
we simply check to see if the block has any data.  If it does not, we
simply proceed without attempting to copy the block contents.